+++ python/GafferSceneUI/__init__.py
@@ -226,6 +226,7 @@ from . import RandomPrimitiveVariableUI
 from . import PointInstancerCoreUI
 from . import PointInstancerUI
 from . import VisibilityQueryUI
+from . import PrimitiveVariableTypeUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.