+++ Changes.md
@@ -5,6 +5,7 @@ Features
 --------
 
 - VisibilityQuery : Added a node for querying the visibility of a particular scene location.
+- PrimitiveVariableType : Added a node for changing the type and interpretation of primitive variables (#7111).
 
 Improvements
 ------------